FeatureDescriptors

FeatureDescriptors.Descriptor — Type

Descriptor

The Descriptors interface lets users define the properties and relationships between the features being used in their models or other data-intensive workflows.

Descriptor supertypes define some general properties that can be overloaded for specific subtypes as required, such as the tables to load or the quantitative variable to use.

Users may define their own Descriptor` type system for their own purposes but must extend the following methods:

sources: The sources for the data needed to build the Desriptor feature.

quantity_key: The quantitative variable for the Descriptor.

categorical_keys: The categorical variables for the Descriptor.

FeatureDescriptors.categorical_keys — Function

categorical_keys(::Descriptor) -> Vector

Returns the categorical variables for the Descriptor, e.g. [:colour, :shape]. Returns an empty vector if no categorical variables exist.

FeatureDescriptors.label — Method

label(D::Descriptor) -> Symbol

Returns a symbol representation of the Descriptor.

FeatureDescriptors.quantity_key — Function

quantity_key(::Descriptor) -> Symbol

Returns the quantitative variable for the Descriptor, e.g. :temperature.

FeatureDescriptors.sources — Function

sources(::Descriptor) -> Vector{String}

Returns the data sources for the Descriptor, e.g. the names or paths to the tables containing the required data: ["temperature.csv", "locations.csv"].

TestUtils

FeatureDescriptors.TestUtils — Module

FeatureDescriptors.TestUtils

Provides a fake Descriptor for testing purposes and a test_interface function to check that any Descriptor subtype implements the expected API.

FeatureDescriptors.TestUtils.FakeDescriptor — Type

FakeDescriptor <: Descriptor

A fake Descriptor for testing purposes only.

FeatureDescriptors.TestUtils.test_interface — Method

TestUtils.test_interface(D::Type{<:Descriptor})

Test that a subtype of Descriptor implements the expected API.

Index

FeatureDescriptors.TestUtils

FeatureDescriptors.Descriptor

FeatureDescriptors.TestUtils.FakeDescriptor

FeatureDescriptors.TestUtils.test_interface

FeatureDescriptors.categorical_keys

FeatureDescriptors.label

FeatureDescriptors.quantity_key

FeatureDescriptors.sources